Added a Level Generator to Create Game Entities on Top of the Map
This commit adds a level generator which fetches reference to the map generator. Instead of the map generator recreating the data on its own (which would leave other entities in the level with no map to interact with) the level generator will handle everything "on top" of the map and call the map generator's functions to regenerate the data, as appropriate.